requirements.txt
# Install these requirements if you wish to contribute to this project.

# ===========================================================================
# Packages essential to this app. Needed by anyone who wants to use this app.
# ===========================================================================
django==1.4.3
django-cms==2.3.5
# There is a problem with STATIC_FILES in the latest commit (as of 2012-02)
-e git+git://github.com/fivethreeo/cmsplugin-blog.git@cd91a4eb066437aa254574759eed51da973314ab#egg=cmsplugin_blog
simple-translation==0.8.6
South==0.7.6


# ==============================================================
# Packages needed for running the tests. Needed by contributors.
# ==============================================================
Fabric
django-libs
factory-boy
mock
django-nose
coverage
django-coverage
ipython
ipdb
